tests/wpt/tests/css/css-color/parsing/color-valid-contrast-color-function.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>Test CSS Color Module 5 contrast-color() parsing valid values</title>
<link rel="author" title="Sam Weinig" href="mailto:sam@webkit.org">
<link rel="help" href="https://drafts.csswg.org/css-color-5/#contrast-color">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
<style>
  #container {
    container-type: inline-size;
    width: 10px;
    color: pink;
  }
</style>
</head>
<body>
<div id="container">
  <div id="target"></div>
</div>
<script>

// Test basic legacy colors
test_valid_value("background-color", "contrast-color(white)");
test_valid_value("background-color", "contrast-color(black)");
test_valid_value("background-color", "contrast-color(pink)");

// Test non-legacy colors
test_valid_value("background-color", "contrast-color(color(srgb 1 0 1 / 0.5))");
test_valid_value("background-color", "contrast-color(lab(0.2 0.5 0.2))");

// Test out-of-gamut colors
test_valid_value("background-color", "contrast-color(color(srgb 10 10 10))");
test_valid_value("background-color", "contrast-color(color(srgb -10 -10 -10))");

// Test nested contrast-color
test_valid_value("background-color", "contrast-color(contrast-color(pink))");

// Test currentColor
test_valid_value("background-color", "contrast-color(currentcolor)");

// Test color using calc().
test_valid_value("background-color", "contrast-color(color(srgb calc(0.5) calc(1 + (sign(20cqw - 10px) * 0.5)) 1 / .5))", "contrast-color(color(srgb calc(0.5) calc(1 + (0.5 * sign(20cqw - 10px))) 1 / 0.5))");

</script>
</body>
</html>
